refactor(Parser): 커스텀 구분자 사용 여부에 따른 파싱 전략 분리
