+++ src/main/java/calculator/model/ParsingStrategy.java
@@ -1,0 +1,7 @@
+package calculator.model;
+
+import java.util.List;
+
+public interface ParsingStrategy {
+    List<Integer> parse(String input);
+}
